Move Array tests from test_libtiledb to test_array
